pyproject.toml
# event/pyproject.toml

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openstarlab_event"
version = "0.1.15"
description = "openstarlab event modeliing package"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name="Calvin Yeung", email="yeung.chikwong@g.sp.m.is.nagoya-u.ac.jp" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "torchprofile",
    "optuna",
    "matplotlib",
    "matplotsoccer",
    "scipy",
]

license = { text = "Apache License 2.0" }

[tool.setuptools.packages.find]
where = ["."]

[project.urls]
Homepage = "https://github.com/open-starlab/Event"